# Переработан механизм позднего связывания, сняты ограничения на имена классов и пространств имен. # Ограничения на вставку параметров с точкой. # Исключение одинаковых типов экспериментов. # При загрузке модуля происходит проверка, модуль ли это. # Добавлен фильтр типов файлов при добавлении модуля. # Работает Файл->Выход.

+++ micro4-1/micro4-1/AlphaParticle.cs
@@ -52,10 +52,10 @@ namespace micro4_1
             ay = fy / 6.7f;
 
             this.vy += ay;
-            this.y -= this.vy;
+            this.y -= (float)this.vy;
 
             this.vx += ax;
-            this.x += this.vx;
+            this.x += (float)this.vx;
         }
         #endregion
